fix(postgres): handle null and score-type issues in request logs and full-text search
- request_log_stats: COALESCE success_rate so an empty table returns 0
  instead of NULL (fixes 'unexpected null' 500 on the postgres backend)
- request_logs DTO: avg_latency_ms is Option<f64>; reconcile handler
- search_fulltext: ts_rank returns FLOAT4, decode score as f32 directly
  instead of f64 (fixes FLOAT8/FLOAT4 mismatch)

+++ crates/memayu-storage-postgres/src/provider.rs
@@ -183,11 +183,13 @@ impl StorageProvider for PostgresProvider {
 
         let mut out = Vec::with_capacity(rows.len());
         for row in rows {
-            let score: f64 = row
+            // ts_rank returns `real` (FLOAT4), so decode directly as f32 to
+            // match the StorageProvider contract instead of a FLOAT8 cast.
+            let score: f32 = row
                 .try_get("score")
                 .map_err(|e| StorageError::Other(format!("read full-text score: {e}")))?;
             let mem = memory_from_row(&row)?;
-            out.push((mem, score as f32));
+            out.push((mem, score));
         }
         Ok(out)
     }
